requirements.txt
streamlit
pytesseract
pillow
opencv-python
numpy
python-dotenv
requests
transformers
torch
huggingface_hub>=0.23.0
pdf2image

python-docx
sentence-transformers
faiss-cpu  
pymupdf
pandas 
PyPDF2